Refactor tests and included tests for new backends and processors
