+++ tests/paz/backend/image/opencv_image_test.py
@@ -3,17 +3,8 @@ import os
 from tensorflow.keras.utils import get_file
 import pytest
 
-from paz.backend.image import opencv_image
-from paz.backend.image.opencv_image import split_and_normalize_alpha_channel
-
-# TODO:
-# Add tests for the following random functions:
-# random_saturation
-# random_brightness
-# random_contrast
-# random_hue
-# random_plain_background
-# show_image
+from paz.backend import image as opencv_image
+from paz.backend.image import split_and_normalize_alpha_channel
 
 
 @pytest.fixture
