+++ tests/paz/optimization/losses/loss_test.py
